+++ docs/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 3.3.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Added documentation and tests for retrieving instance information after import (#)
 
 
 3.3.1 (2023-09-14)